refactor package declaration
